Example of how to do coloring in S3D
Updated attributions.md with link to spaceship model used in video.
Added Crane_S3D_Settings.zip with settings used in the video to help as a quick start. This may not be the right place for this.
Updated user-submitted-guides.md with youtube link to my video for setting up a model in S3D.